Limit content script to owned websites only
The existing content script injects anywhere, even though the functionality is
only useful to a limited set of websites. Replaced wildcard '*' with a specific
list of domains to prevent conflicts between add-ons that are based on the same
code.

Also removed the obsolete comment about adding domains, this is no longer
possible.